+++ apis/database/v1alpha1/zz_mysql_types.go
@@ -25,14 +25,14 @@ type MySQLInitParameters struct {
 
 	// The number of Linode Instance nodes deployed to the Managed Database. (default 1)
 	// The number of Linode Instance nodes deployed to the Managed Database. Defaults to 1.
-	ClusterSize *int64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
+	ClusterSize *float64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
 
 	// Whether the Managed Databases is encrypted. (default false)
 	// Whether the Managed Databases is encrypted.
 	Encrypted *bool `json:"encrypted,omitempty" tf:"encrypted,omitempty"`
 
 	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
-	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.26)
+	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
 	EngineID *string `json:"engineId,omitempty" tf:"engine_id,omitempty"`
 
 	// A unique, user-defined string referring to the Managed Database.
@@ -67,7 +67,7 @@ type MySQLObservation struct {
 
 	// The number of Linode Instance nodes deployed to the Managed Database. (default 1)
 	// The number of Linode Instance nodes deployed to the Managed Database. Defaults to 1.
-	ClusterSize *int64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
+	ClusterSize *float64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
 
 	// When this Managed Database was created.
 	// When this Managed Database was created.
@@ -82,7 +82,7 @@ type MySQLObservation struct {
 	Engine *string `json:"engine,omitempty" tf:"engine,omitempty"`
 
 	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
-	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.26)
+	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
 	EngineID *string `json:"engineId,omitempty" tf:"engine_id,omitempty"`
 
 	// The primary host for the Managed Database.
@@ -142,7 +142,7 @@ type MySQLParameters struct {
 	// The number of Linode Instance nodes deployed to the Managed Database. (default 1)
 	// The number of Linode Instance nodes deployed to the Managed Database. Defaults to 1.
 	// +kubebuilder:validation:Optional
-	ClusterSize *int64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
+	ClusterSize *float64 `json:"clusterSize,omitempty" tf:"cluster_size,omitempty"`
 
 	// Whether the Managed Databases is encrypted. (default false)
 	// Whether the Managed Databases is encrypted.
@@ -150,7 +150,7 @@ type MySQLParameters struct {
 	Encrypted *bool `json:"encrypted,omitempty" tf:"encrypted,omitempty"`
 
 	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
-	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.26)
+	// The Managed Database engine in engine/version format. (e.g. mysql/8.0.30)
 	// +kubebuilder:validation:Optional
 	EngineID *string `json:"engineId,omitempty" tf:"engine_id,omitempty"`
 
@@ -192,7 +192,7 @@ type UpdatesInitParameters struct {
 
 	// The maximum maintenance window time in hours. (1..3)
 	// The maximum maintenance window time in hours.
-	Duration *int64 `json:"duration,omitempty" tf:"duration,omitempty"`
+	Duration *float64 `json:"duration,omitempty" tf:"duration,omitempty"`
 
 	// Whether maintenance occurs on a weekly or monthly basis. (weekly, monthly)
 	// Whether maintenance occurs on a weekly or monthly basis.
@@ -200,11 +200,11 @@ type UpdatesInitParameters struct {
 
 	// The hour to begin maintenance based in UTC time. (0..23)
 	// The hour to begin maintenance based in UTC time.
-	HourOfDay *int64 `json:"hourOfDay,omitempty" tf:"hour_of_day,omitempty"`
+	HourOfDay *float64 `json:"hourOfDay,omitempty" tf:"hour_of_day,omitempty"`
 
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates. (1..4)
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates.
-	WeekOfMonth *int64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
+	WeekOfMonth *float64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
 }
 
 type UpdatesObservation struct {
@@ -215,7 +215,7 @@ type UpdatesObservation struct {
 
 	// The maximum maintenance window time in hours. (1..3)
 	// The maximum maintenance window time in hours.
-	Duration *int64 `json:"duration,omitempty" tf:"duration,omitempty"`
+	Duration *float64 `json:"duration,omitempty" tf:"duration,omitempty"`
 
 	// Whether maintenance occurs on a weekly or monthly basis. (weekly, monthly)
 	// Whether maintenance occurs on a weekly or monthly basis.
@@ -223,11 +223,11 @@ type UpdatesObservation struct {
 
 	// The hour to begin maintenance based in UTC time. (0..23)
 	// The hour to begin maintenance based in UTC time.
-	HourOfDay *int64 `json:"hourOfDay,omitempty" tf:"hour_of_day,omitempty"`
+	HourOfDay *float64 `json:"hourOfDay,omitempty" tf:"hour_of_day,omitempty"`
 
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates. (1..4)
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates.
-	WeekOfMonth *int64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
+	WeekOfMonth *float64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
 }
 
 type UpdatesParameters struct {
@@ -240,7 +240,7 @@ type UpdatesParameters struct {
 	// The maximum maintenance window time in hours. (1..3)
 	// The maximum maintenance window time in hours.
 	// +kubebuilder:validation:Optional
-	Duration *int64 `json:"duration" tf:"duration,omitempty"`
+	Duration *float64 `json:"duration" tf:"duration,omitempty"`
 
 	// Whether maintenance occurs on a weekly or monthly basis. (weekly, monthly)
 	// Whether maintenance occurs on a weekly or monthly basis.
@@ -250,12 +250,12 @@ type UpdatesParameters struct {
 	// The hour to begin maintenance based in UTC time. (0..23)
 	// The hour to begin maintenance based in UTC time.
 	// +kubebuilder:validation:Optional
-	HourOfDay *int64 `json:"hourOfDay" tf:"hour_of_day,omitempty"`
+	HourOfDay *float64 `json:"hourOfDay" tf:"hour_of_day,omitempty"`
 
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates. (1..4)
 	// The week of the month to perform monthly frequency updates. Required for monthly frequency updates.
 	// +kubebuilder:validation:Optional
-	WeekOfMonth *int64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
+	WeekOfMonth *float64 `json:"weekOfMonth,omitempty" tf:"week_of_month,omitempty"`
 }
 
 // MySQLSpec defines the desired state of MySQL
